feat: setup Example unit tests environment
