+++ DMSGoPGP/Tests/DMSGoPGPTests.swift
@@ -1,0 +1,16 @@
+//
+//  DMSGoPGPTests.swift
+//  Pods-DMSGoPGP
+//
+//  Created by Cirno MainasuK on 2019-8-22.
+//
+
+import XCTest
+
+class DMSGoPGPTests: XCTestCase {
+
+    func testSmoke() {
+        
+    }
+
+}
