+++ Example/DMSGoPGP-Example.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
