+++ Example/DMSGoPGP-Example/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  DMSGoPGP-Example
+//
+//  Created by Cirno MainasuK on 2019-8-22.
+//  Copyright © 2019 dimension. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
