+++ Example/DMSGoPGP-ExampleTests/DMSGoPGP_ExampleTests.swift
@@ -1,0 +1,34 @@
+//
+//  DMSGoPGP_ExampleTests.swift
+//  DMSGoPGP-ExampleTests
+//
+//  Created by Cirno MainasuK on 2019-8-22.
+//  Copyright © 2019 dimension. All rights reserved.
+//
+
+import XCTest
+@testable import DMSGoPGP_Example
+
+class DMSGoPGP_ExampleTests: XCTestCase {
+
+    override func setUp() {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
