--- Example/DMSGopenpgp/ViewController.swift
@@ -1,26 +1,0 @@
-//
-//  ViewController.swift
-//  DMSGopenpgp
-//
-//  Created by jk234ert on 08/01/2019.
-//  Copyright (c) 2019 jk234ert. All rights reserved.
-//
-
-import UIKit
-import DMSGoPGP
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-}
-
